fix: resolve SuperFX game boot issues and GH pages deployment
1. Set GitHub pages public-url to /RustySNES/ to fix 404s.
2. Fix SuperFX game header SRAM parsing so games correctly get 32 or 64 KiB RAM.
3. Suppress missing docs warnings in test harness tests.
4. Fix GSU sign extension bug.

+++ crates/rustysnes-cart/src/coproc/gsu.rs
@@ -993,9 +993,9 @@ impl Gsu {
         let operand: i32 = if self.alt2() {
             n as i32
         } else {
-            i32::from(self.r[n])
+            self.r[n] as i16 as i32
         };
-        let sr = i32::from(self.sr());
+        let sr = self.sr() as i16 as i32;
         let carry = if self.alt1() {
             i32::from(self.flag(SFR_CY))
         } else {
@@ -1013,11 +1013,11 @@ impl Gsu {
     // $60-$6f sub/sbc/cmp
     fn i_sub_sbc_cmp(&mut self, n: usize, _mem: &mut GsuMem) {
         let operand: i32 = if !self.alt2() || self.alt1() {
-            i32::from(self.r[n])
+            self.r[n] as i16 as i32
         } else {
             n as i32
         };
-        let sr = i32::from(self.sr());
+        let sr = self.sr() as i16 as i32;
         let borrow = if !self.alt2() && self.alt1() {
             i32::from(!self.flag(SFR_CY))
         } else {
